+++ src/app/Components/ComponentsSamples/BottomSheets/Sheets/SimpleBottomSheet.xaml.cs
@@ -6,15 +6,5 @@ namespace Components.ComponentsSamples.BottomSheets.Sheets
         {
             InitializeComponent();
         }
-
-        private void FirstBottomSheetView_OnWillClose(object sender, EventArgs e)
-        {
-            
-        }
-
-        private void FirstBottomSheetView_OnDidClose(object sender, EventArgs e)
-        {
-            
-        }
     }
 }