+++ CHANGELOG.md
@@ -1,8 +1,19 @@
-## [19.12.0]
+## [20.1.0]
 - Renamed 'IsEllipsized' property to 'IsTruncated' in Label
 - Now checks if the Label is truncated, does not matter if LineBreakMode is set (Ellipsis is apparent)
 - Added new components: SingleLineInputField and MultiLineInputField
 - Can now disable default padding in dui:Entry and dui:Editor
+
+## [20.0.0]
+- [BottomSheet] Will now respond to property changes when its open.
+- [BottomSheet] Added new `Position` property which will determine what position the bottom sheet will present in when opened.
+- [BottomSheet] People can drag the bottom sheet to large when the position is medium and vica versa.
+- [BottomSheet] Consumer can open multiple bottom sheets and close them by using `BottomSheetService.CloseAll()`
+- [BottomSheet][BreakingChange] Replaced `OpenBottomSheet()` with `Open()`
+- [BottomSheet][BreakingChange] ShouldFitContent is replaced with `Position` = `Fit`
+
+## [19.11.1]
+- [ContextMenu] Group `ContextMenuGroup` if at least one item alongside it. 
 
 ## [19.11.0]
 - [SearchBar] Made it possible to Focus and Unfocus